refactor(server): move tools to nodes/tools/*.node.ts; update imports in tests and templates (#211)
