+++ apps/server/__tests__/mcp.provision.dynamic.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { LocalMCPServer } from '../src/mcp/localMcpServer';
+import { LocalMCPServer } from '../src/nodes/mcp/local/local-mcp-server.node';
 import type { McpServerConfig, McpTool } from '../src/mcp/types';
 
 class MockLogger { info=vi.fn(); debug=vi.fn(); error=vi.fn(); }
