+++ apps/server/__tests__/tools/shell.schema.validation.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { ShellToolStaticConfigSchema } from '../../src/tools/shell_command';
+import { ShellToolStaticConfigSchema } from '../../src/nodes/tools/shell-tool.node';
 
 describe('ShellToolStaticConfigSchema validation', () => {
   it('accepts 0 for both timeouts', () => {
