+++ apps/server/src/nodes/tools/finish-tool.node.ts
