+++ apps/server/src/nodes/tools/memory-tool.node.ts
