+++ apps/server/src/nodes/tools/shell-tool.node.ts
