--- apps/server/src/tools/terminateResponse.ts
@@ -1,3 +1,0 @@
-export class TerminateResponse {
-  constructor(public message?: string) {}
-}
